Update README.md
Vamos a jug at en el Sol. 

The Sol release, a single file archive with easy download, for wider audiences. 

I guess Prince was right, “we’re going down, down, down… Cause that’s the only way, to make this cruel cruel world hear what we got to say.” By down we mean DOWNLOADS for the wider audience. Not everyone is a geek.