Find this Sharp Aquos…
I cannot find this model on Sharps website. The  bezel says Sharp. They have similar model numbers but not this one. 

Firmware for similar models isn’t even detected. I know it takes firmware, the menus have an option for upgrading it and the dumb-TV does really intelligent things. Like based on my conversation or what’s on the screen.