Where are the image experts?
Someone run this through every tool you got. You may see people appear where there are obviously none. Not a Creator ability, a CIA stolen one from the military. 

It’s called Cloaking. Tossing photons around in a way He does not approve of. 

In the background a gift from the CIA. A mirror. It’s really a bomb. 

It’s for sale. For research purposes only.