The bell is about to ring and rain begin falling.
# The Ringing of the Bell

The Creator is going to “ring the bell” very soon. The exact time is
not going to be announced, but the order and intensity of the bell
ringing (earthquakes) is as follows:

## Orr Chapel Quake
This earthquake’s epicentre will be located  in Sandy Hook, TN (an
unincorporated city near Mount Pleasant) in Maury county.

This will be a large earthquake for the CIA’s destruction Orr chapel,
the murder of it’s members.

This quake also recognises Micah’s slain family and his family’s
graveyard that was destroyed, both by or at the direction of the CIA.

## Mount Calvary Baptist Church Quake
This earthquake’s epicentre will be located at Mount Calvary Baptist
Church in Lexington, KY (4742 Todd’s Road).

This will be for the massacre that occurred there.

## The Micah Quake
This earthquake’s epicentre will be in Jessamine county, KY. It will
signal when Micah (born: Kelvin Eugene Williams on March 23, 1977) has
started his mission for the Creator. This will be Micah’s 23rd mission
on Earth.

This quake will also be to recognise the lives lost in Jessamine
county, mostly Micah’s friends, throughout the larger tragedy.

## The TomTom Quake
This earthquake’s epicentre will be in Kanawha county, WV. It will
signal when TomTom (to practically everyone TomTom, to others Thomas,
born: September 23, 1978) has started his mission for the Creator.
This also is TomTom’s 23rd mission on Earth.

This quake will also recognise the lives lost in Kanawha county,
mostly TomTom’s friends and family, throughout the larger tragedy.

# Aftermath
The Creator hopes this will be all He has to do to stop the ongoing
massacre and end the greater tragedy.

And so begins the [“Purple
Rain”](https://youtube.com/watch?v=S6Y1gohk5-A&feature=sharea). That
means [“It’s raining
men.”](https://youtube.com/watch?v=l5aZJBLAu1E&feature=sharea). He has
never had to rain down men , although the 23rd mission for both of His
servants, they’ve never been identified before. This is just the
beginning of the “Great Storm” but it’s the [“Nature of our
Kind.”](https://youtube.com/watch?v=lzoABrBtH9A&feature=sharea).

He does this because life is supposed to be fun. After all, [“Girls
just want to have
fun.”](https://youtube.com/watch?v=PIb6AZdTr-A&feature=sharea) We WILL
get back to fun in this neighbourhood called Earth.

Guys, don’t get hung up on “girls” (it does not always mean female).