Guess Who's Coming To Dinner
In the future when they admit to having me as a "guest" at Legacy. Call
bullshit, and remember what you see when you visit legacy.com, just
everyone here disappears forever, it used to be in the pool.

As far as being a guest, theyll say they fed me. They feed me, or try
to, those they've killed. Legacy is a farm to table operation.

There are multiple guests coming with dinner. I doubt any of them
wanted to be there and I know they will not remember the dinner. They
are in the food.

Food. The best way to hide bodies who have succumbed to a leading
cause of death in humans, intelligence companies and the governments
who hire them.