This commit storm NOT brought to you by Primus Labs
"At Primus Labs we can help you hide the murder of your fellow citizens
or your biggest client's (before the pharmaceutical industry) citizens
with nice, pretty, organic (and we will help keep them believing that
means it grew in a field)."

Cheryl was the cool mom across the street where I grew up. Ive never
heard of this brand. There's no website. The Creator assured me that
was who was in the food delivered to me by the CIA told it came from
the food bank. It probably did. They like to give back to the
communities they serve...I mean took from.

There is sometimes a lot of truth in satire. For mine, there is always
#PrayToVerify

Subtle reminder/instructions on #PrayToVerify or #PTV. Pray to the God
you were first taught to, the way you were taught to, and ask "Is this
true?" See how you feel about it then.