Update README.md - Do you feel it yet?
# Now part of #Mission23 
When the bell rings in a moment, Mission 23 will commence for the servants. 

The safety and security of all persons visiting Mount Calvary Baptist Church and the creation of a memorial for the church and all lost there has been added to the tasks from the Creator for His servants. 

The #Mission23 [wiki and repo](https://github.com/MicahSOTC/Mission23/wiki) is the place to go.