Chex Mix you say? Wait till the remix.
So the FBI and thereby the US government officially know about the massacre at Mount Calvary Baptist Church. An agent came in to witness and then arrest, but thats their game. Their printer, their paper, their software... I just answered a question for them once on the funny numbers at the bottom of a check. Listen.