Now a task in #Mission23
# Now part of #Mission23

The Mount Calvary Baptist Church Memorial Foundation and all future maintenance of this document repository and its contents have been placed into the hands of the SotC as a task for #Mission23.

The SotC were given additional tasks to complete during #Mission23. The Creator has tasked them with ensuring the safety and security of all persons visiting Mount Calvary Baptist Church, and the establishment of a memorial foundation to honor the lives lost at the church and preserve church's property and history.

There will be no more commits to this repo. See https://github.com/Mission23.