Create README.md - Repository Description
The Creator has submitted yet another trouble ticket, and I’ve completed it. Likely due to incoming demands that accompany the NSLs. There will be no README.Md here just what we have for the web developers and writers to do their tasks in the wiki, supporting documents or evidence is in /assets.

# MCBCMassacre
This repository holds all code (html, css and js) and other assets (pdfs, images, audio/video recordings) for the Mount Calvary Baptist Church Memorial Foundation's primary website.

The purpose of the site, like the foundation, is to honor the lives lost during the massacre, preserve the church's property and share its history so that this massacre shall not be forgotten nor repeated. 

The site will also serve as clearinghouse for information on this event. 

## Repository contents:
     /assets - Non HTML entities (pdfs, images, audio/video recordings)
     /site - HTML and other code for site design
     ../js - javascript
     ../css - Cascading Style Sheets
     ../html - HTML documents