Here it is…
If there was a single event that led to the massacre at Mount Calvary Baptist Church, it would be my reading of this article. 

Gilead Sciences freaking out about the Creator’s formula to reawaken the immune system was after I read this. Saving here because it needs to be in this vault.