Updated README.md
The Creator is going to “ring the bell” very soon. The exact time is
not going to be announced.  Scientists have often said that a very
strong earthquake will ring Earth like a bell.  These earthquakes are
intended to _shine a light_ on a horrible, unnecessary tragedy that
culminated with a [massacre of a church's
congregation](https://github.com/mission23/mission23/wiki/The-Massacre-
at-Mount-Calvary-Baptist-Church) during the height of their worship
hour.

Regardless of how much work folks have done to get the story out and
reported for your safety, the federal government and federal law
enforcement agencies are still putting pressure on the Commonwealth of
Kentucky's attorney and law enforcement to turn a blind eye and IGNORE
the actions of the CIA. The CIA continues to kill in the very place
that people go to and expect to be the safest and closest to God, who
is none other than, the Creator (my pissed off boss).

Obviously drunk on power, the CIA, the federal government and federal
law enforcement need to sober up quickly, so the Creator, has decided
to ring the bell.  It's a warning that they have forgotten who's
universe, Earth, and country this landmass really is.

In addition to showing how displeased He is, the Creator is also
honoring lives that have cut short in response to a set of "cures"
(some of objectives in #mission23) that should have let all of these
lives go on for far more years than 99% of the human population
believe is even possible.  The way He designed and wants them to be
lived.

It also signals the start of #mission23, and says "Thanks" to everyone
who has tolerated the federal government's pressure when they knew the
right thing to do and they simply wanted to do it.

The order and intensity of the bell ringing (earthquakes) goes as
follows:

## Orr Chapel Quake
This earthquake’s epicentre will be located  in Sandy Hook, TN (the
former unincorporated city near, now neighborhood of Mount Pleasant)
in Maury county. The Rogers family farm.

This will be a large earthquake for the CIA’s destruction of Orr
Chapel ([Google Maps](https://goo.gl/maps/XMMdNdpGjU3SMMKQ8)) and the
murder of all of it’s identified members.

This quake also recognises Micah’s slain family and his family’s
graveyard that was destroyed, both by or at the direction of the CIA.

## Mount Calvary Baptist Church Quake
This earthquake’s epicentre will be located at Mount Calvary Baptist
Church in Lexington, KY (4742 Todds Road).

This will be the largest quake for the massacre that occurred there.

See: [The Massacre at Mount Calvary Baptist
Church](https://github.com/Mission23/MCBCMassacre/wiki/Massacre-at-Moun
t-Calvary-Baptist-Church)

## The Micah Quake
This earthquake’s epicentre will be in Jessamine county, KY. It will
signal when Micah (born: Kelvin Eugene Williams on March 23, 1977) has
started his mission for the Creator. This will be Micah’s 23rd mission
on Earth.

This minor quake will also be to recognize the lives lost in Jessamine
county, mostly Micah’s friends, throughout the larger tragedy.

## The Thomas Quake
This earthquake’s epicentre will be in Kanawha county (Hughes Creek),
WV. It will signal when TomTom (to practically everyone TomTom, to
others Thomas, born: September 23, 1978) has started his mission for
the Creator. This also is TomTom’s 23rd mission on Earth.

This minor quake will also recognize the lives lost in Kanawha county,
mostly TomTom’s friends and family, throughout the larger tragedy.

# Press Conferences
These are tentative.  Conditions on the ground are very fluid--and I
don't mean from the quakes.  Please check back here for updates to the
schedule.

## Orr Chapel
At high noon the day immediately following the shaking under our feet
the SotC will have a press conference at the site of Orr Chapel &
Sheepneck cemeteries. Located along West Sheepneck Road in Sandy Hook,
TN.
## Mount Calvary Baptist Church
On the second day following the shaking, the SotC will be at Mount
Calvary Baptist Church at 4742 Todds Road in Lexington, KY at high
noon.

# Suggested Listening
## The Sound and the Fury
By Vandaveer - [Listen on YouTube
Music](https://music.youtube.com/watch?v=OpLeRY6NIhA&si=cmrf2QAr0f544FM
Z)

## 1999
By Prince
[Listen on YouTube
Music](https://music.youtube.com/watch?v=rblt2EtFfC4&si=yghTVWBWzrB5KjD
7)