The plan...
Update 2023-10-08 15:47
According to Danny, the apparent plan is to lock me out in the cold,
he has my keys and my coats have disappeared. If I were to break the
window, he will cut me up as this is his apartment.

Funny thing is, I have video of him breaking that very window and
audio of the Riverdale PD telling me he could.