about being on HIS side...
To clarify the last commit.

The local law enforcement are on Danny's side. They routinely stand in
the room and watch him try to kill me.

They just tried to kill me for this commit.

The Creator is DONE with Clayton County. Atlanta's human trafficking
problem is headquartered here in Clayton county. The home of
Hartsfield Jackson Atlanta International airport.