P3 Tips / Crimestoppers
Communities whose Crime stopping programs or local law enforcement agencies which utilize any of the P3 apps should stop using it immediately!

It is managed by the NSA in partnership with the CIA to HIDE activities they cannot be discovered by law enforcement, you know, those who actually protect and serve humanity. 

The Creator has now added to mission23 a task to replace all of the services and apps that people depend on that are owned and/or operated by humanity whose owners are hidden or obfuscated because they are using the data for their own nefarious purposes. More coming in the wiki. He knows all the tricks the devil plays, and that means more work for us. Yay!

Dear LEAs: Don’t think we are gonna do it all!!! We are gonna lean on y’all for assistance. Y’all get paychecks, we don’t. I’m serious, He don’t even have a checkbook.