Update README.md - commits & comments
# commits are everything. 
each file added or changed is done so via a commit. you can view the description of each and comment on them. a github.com account is required. 

warning: this is still an active situation with the government killing those who are involved or looking into it. 

i recommend using a separate anonymized account for commenting right now. -m