Update README.md - Complete Redo (Ctt)
# When in doubt, pray  
The following summary is very hard to believe. We ask that any reader who doubts any part of it to pray and ask for guidance. 
***
The following is a very brief summary of The Massacre at Mount Calvary Baptist Church on Todds Road in Lexington, KY. For a more detailed write up [click here](https://github.com/mission23/mission23/wiki/The-Massacre-at-Mount-Calvary-Baptist-Church).
***
On Sunday April 2nd, 2023 a CIA team led by Annie Jacobsen stormed Mount Calvary Baptist Church in Lexington, Kentucky near the end of the morning worship services. 

Prior to entering the church the CIA deployed and activated cellular network equipment which would cause all cell phones in the nearby area to automatically connect to it (a “cell on wheels”). 

After blocking the back two exits, about fifty men and a couple men gathered in the vestibule at the front entrance of the church.  A projection (like a movie) began to show on all of the windows of the sanctuary that made it seem as if escaping through a window would be impossible. The group in the vestibule then burst in announcing it was a law enforcement action. 

With all exits blocked and cellphones unable to call actual 911 due to the setup of the cellular equipment the CIA brought, the invading force began the systematic murder of everyone inside. 

The Rev. Dr. Felix G. Williams, III was tortured extensively by several in the group who demanded to know about church operations and his family relationships and finally murdered. 

The CIA then setup a charade church called Mount Calvary Missionary Baptist Church, Inc. and began producing videos to portray normalcy at the church using actors who resembled pastors and members of the church. The CIA has opened the church week after week in an effort to murder any remaining members or those with a relationship with the church that show up. 

The CIA also began holding funerals to eliminate any in the surrounding communities who expressed interest in those previously killed the church. 
***
A more detailed write up of the event and subsequent activities, including things you should do to stay safe can be found on the [mission23 wiki](https://github.com/mission23/mission23/wiki/The-Massacre-at-Mount-Calvary-Baptist-Church).
***
### Warning
No one should call or visit Mount Calvary Baptist Church or a church by any other name on Todds Road until they have seen this coverage on the local and national news channels. 

It is recommended that any person who feels the need to report a person missing that was known to be going to a church on Todds Road to delay making the report due a government cover up. The CIA is actively identifying and killing anyone who makes a report of a missing person like the one aforementioned. 

This repository is focused on Mount Calvary Baptist Church on Todds Road in Lexington due to its maintainers long relationship with the church and family relationship with the pastor. Other churches on Todds Road we cannot speak for, but fear they have become part of the massacre turned genocide and cover up by the US government and CIA.