Update current_status.txt
Im writing articles to get more eyes on this repo and wiki. 

you can read everything at: https://medium.com/@micah_92788

they ARE trying to sey up an arrest. but the ring is in the bank! and the phone has the longest password ever and wont work with any other face.