Create 23andMe.md
# 23AndMe
Currently, I am not authorizing 23andMe to unlock or grant any access to the kelvin@kelvinwilliams.net account. 

Please do not open it even if an ID with my name on it is presented with my name on it or emails exchanged with my email address unless this has been updated. 

Also, please do not delete or change the names on the kits on file.