GATech Warning!!! Direct from the Creator.
# Georgia Tech Warning
The CIA is trying to tie up lose ends, meaning killing anyone that may
know things or the true GATech Alumni (2013) Kelvin Eugene Williams.

## Cloaking
The CIA utilizes a cloaking technology that will render a person
practically invisible to the naked eye. It’s stolen from the military.

You should be able to find some documentation on it by searching. I
found a [this
article](https://www.nature.com/articles/s41598-017-08505-w) on photon
entanglement and cloaking that told me a lot.

It’s an electromagnetic field, I know there’s equipment on campus that
can and does detect it. Remember this is a revenge of the nerds and
you are your brother’s and sister’s keeper.

They CAN and DO use a companion technology that enables them to pass
through walls and doors.

The Creator says they are on campus tonight. (To the nerd reading
this: START RINGING THE FIREBELL! Let everyone on campus know, I bet
you have a way.)

Everyone on campus needs should not sleep alone. Pair up with another
student. Four hours of sleep is all you need. While one is sleeping,
the other is up watching and listening.

(Saving. Reload in a couple minutes for more.)